@@ -37,8 +37,7 @@
     "extends": "graylog"
   },
   "dependencies": {
-    "@babel/preset-env": "7.16.11",
-    "@babel/preset-typescript": "7.16.7",
+    "babel-preset-graylog": "file:../babel-preset-graylog",
     "create-react-class": "15.7.0",
     "eslint-config-graylog": "file:../eslint-config-graylog",
     "formik": "2.2.9",
@@ -58,15 +57,14 @@
     "react-router-dom": "5.3.0",
     "reflux": "0.2.13",
     "styled-components": "5.3.5",
-    "typescript": "4.7.2",
-    "webpack": "5.72.1",
+    "typescript": "4.7.3",
+    "webpack": "5.73.0",
     "webpack-cleanup-plugin": "0.5.1",
     "webpack-cli": "4.9.2",
     "webpack-merge": "4.2.2"
   },
   "devDependencies": {
     "@babel/cli": "7.16.8",
-    "@babel/core": "7.16.12",
     "babel-plugin-add-module-exports": "^1.0.4"
   }
 }
