@@ -32,6 +32,8 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.util.Arrays;
+import java.util.Collections;
+import java.util.LinkedHashSet;
 import java.util.List;
 
 import static io.restassured.RestAssured.given;
@@ -91,7 +93,7 @@ public class SearchDriver {
                 .build();
         SearchDTO s = SearchDTO.builder()
                 .id(new ObjectId().toHexString())
-                .queries(ImmutableSet.of(q))
+                .queries(q)
                 .build();
 
         return JsonUtils.toJsonString(s);
