@@ -27,6 +27,8 @@ import org.graylog.plugins.views.search.Query;
 import org.graylog.plugins.views.search.Search;
 
 import javax.annotation.Nullable;
+import java.util.Arrays;
+import java.util.LinkedHashSet;
 import java.util.Set;
 import java.util.stream.Collectors;
 
@@ -41,7 +43,7 @@ public abstract class SearchDTO {
     public abstract String id();
 
     @JsonProperty
-    public abstract Set<QueryDTO> queries();
+    public abstract LinkedHashSet<QueryDTO> queries();
 
     @JsonProperty
     public abstract Set<Parameter> parameters();
@@ -53,7 +55,7 @@ public abstract class SearchDTO {
                 .queries(search.queries()
                         .stream()
                         .map(QueryDTO::fromQuery)
-                        .collect(Collectors.toSet()))
+                        .collect(Collectors.toCollection(LinkedHashSet::new)))
                 .build();
     }
 
@@ -69,7 +71,11 @@ public abstract class SearchDTO {
         public abstract String id();
 
         @JsonProperty
-        public abstract Builder queries(Set<QueryDTO> queries);
+        public abstract Builder queries(LinkedHashSet<QueryDTO> queries);
+
+        public Builder queries(QueryDTO... queries) {
+            return this.queries(new LinkedHashSet<>(Arrays.asList(queries)));
+        }
 
         @JsonProperty
         public abstract Builder parameters(Set<Parameter> parameters);
@@ -79,7 +85,7 @@ public abstract class SearchDTO {
         @JsonCreator
         static Builder create() {
             return new AutoValue_SearchDTO.Builder()
-                    .queries(ImmutableSet.of())
+                    .queries(new LinkedHashSet<>())
                     .parameters(of());
         }
     }
