@@ -36,6 +36,7 @@ import org.graylog.scheduler.Job;
 import org.graylog.scheduler.JobDefinitionConfig;
 import org.graylog.scheduler.JobSchedule;
 import org.graylog.scheduler.JobTriggerData;
+import org.graylog.scheduler.capabilities.SchedulerCapabilities;
 import org.graylog.security.authservice.AuthServiceBackend;
 import org.graylog.security.authservice.AuthServiceBackendConfig;
 import org.graylog2.audit.AuditEventType;
@@ -387,8 +388,13 @@ public abstract class PluginModule extends Graylog2Module {
 
 
     protected void addClusterConfigValidator(Class<?> configClass, Class<? extends ClusterConfigValidator> configValidatorClass) {
-
         clusterConfigMapBinder().addBinding(configClass).to(configValidatorClass);
+    }
 
+    protected Multibinder<SchedulerCapabilities> schdulerCapabilitiesBinder() {
+        return Multibinder.newSetBinder(binder(), SchedulerCapabilities.class);
+    }
+    protected void addSchedulerCapabilities(Class<? extends SchedulerCapabilities> schedulerCapabilitiesClass) {
+        schdulerCapabilitiesBinder().addBinding().to(schedulerCapabilitiesClass);
     }
 }
