@@ -15,28 +15,6 @@
  * <http://www.mongodb.com/licensing/server-side-public-license>.
  */
 module.exports = {
-  presets: [
-    // eslint-disable-next-line global-require
-    [require('@babel/preset-env'), {
-      modules: false,
-      useBuiltIns: 'entry',
-      corejs: '3.9',
-      shippedProposals: true,
-    }],
-    '@babel/react',
-    '@babel/typescript',
-  ],
-  plugins: [
-    '@babel/plugin-syntax-dynamic-import',
-    'babel-plugin-styled-components',
-  ],
-  env: {
-    test: {
-      presets: ['@babel/env', '@babel/typescript'],
-      plugins: [
-        'babel-plugin-dynamic-import-node',
-        '@babel/plugin-transform-runtime',
-      ],
-    },
-  },
+  // eslint-disable-next-line global-require
+  presets: [require('babel-preset-graylog')],
 };
