@@ -16,17 +16,15 @@
  */
 import React from 'react';
 
-class NoopRetentionStrategySummary extends React.Component {
-  render() {
-    return (
-      <div>
-        <dl>
-          <dt>Index retention strategy:</dt>
-          <dd>Do nothing</dd>
-        </dl>
-      </div>
-    );
-  }
-}
+const NoopRetentionStrategySummary = () => {
+  return (
+    <div>
+      <dl>
+        <dt>Index retention strategy:</dt>
+        <dd>Do nothing</dd>
+      </dl>
+    </div>
+  );
+};
 
 export default NoopRetentionStrategySummary;
