@@ -19,23 +19,23 @@ import React from 'react';
 
 import NumberUtils from 'util/NumberUtils';
 
-class SizeBasedRotationStrategySummary extends React.Component {
-  static propTypes = {
-    config: PropTypes.object.isRequired,
-  };
+const SizeBasedRotationStrategySummary = ({ config }) => {
+  const { max_size: maxSize } = config;
 
-  render() {
-    return (
-      <div>
-        <dl>
-          <dt>Index rotation strategy:</dt>
-          <dd>Index Size</dd>
-          <dt>Max index size:</dt>
-          <dd>{this.props.config.max_size} bytes ({NumberUtils.formatBytes(this.props.config.max_size)})</dd>
-        </dl>
-      </div>
-    );
-  }
-}
+  return (
+    <div>
+      <dl>
+        <dt>Index rotation strategy:</dt>
+        <dd>Index Size</dd>
+        <dt>Max index size:</dt>
+        <dd>{maxSize} bytes ({NumberUtils.formatBytes(maxSize)})</dd>
+      </dl>
+    </div>
+  );
+};
+
+SizeBasedRotationStrategySummary.propTypes = {
+  config: PropTypes.object.isRequired,
+};
 
 export default SizeBasedRotationStrategySummary;
