@@ -25,6 +25,16 @@ export type QueryId = string;
 
 export type FilterType = Immutable.Map<string, any>;
 
+export type SearchFilter = {
+  type: 'referenced' | 'inlineQueryString',
+  id?: string,
+  title?: string,
+  description?: string
+  queryString: string
+}
+
+export type FiltersType = Immutable.List<SearchFilter>
+
 export type SearchTypeList = Array<SearchType>;
 type InternalBuilderState = Immutable.Map<string, any>;
 
@@ -33,6 +43,7 @@ type InternalState = {
   query: any,
   timerange: any,
   filter?: FilterType,
+  filters?: FiltersType,
   searchTypes: SearchTypeList,
 };
 
@@ -41,6 +52,7 @@ export type QueryJson = {
   query: any,
   timerange: any,
   filter?: FilterType,
+  filters?: FiltersType,
   search_types: any,
 };
 
@@ -127,8 +139,8 @@ const isNullish = (o: any) => (o === null || o === undefined);
 export default class Query {
   private _value: InternalState;
 
-  constructor(id: QueryId, query: any, timerange: any, filter?: FilterType, searchTypes?: SearchTypeList) {
-    this._value = { id, query, timerange, filter, searchTypes };
+  constructor(id: QueryId, query: any, timerange: any, filter?: FilterType, searchTypes?: SearchTypeList, filters?: FiltersType) {
+    this._value = { id, query, timerange, filter, filters, searchTypes };
   }
 
   get id(): QueryId {
@@ -147,13 +159,17 @@ export default class Query {
     return this._value.filter;
   }
 
+  get filters(): FiltersType | null | undefined {
+    return this._value.filters;
+  }
+
   get searchTypes(): SearchTypeList {
     return this._value.searchTypes;
   }
 
   // eslint-disable-next-line @typescript-eslint/no-use-before-define
   toBuilder(): Builder {
-    const { id, query, timerange, filter, searchTypes } = this._value;
+    const { id, query, timerange, filter, filters, searchTypes } = this._value;
 
     // eslint-disable-next-line @typescript-eslint/no-use-before-define
     const builder = Query.builder()
@@ -161,6 +177,7 @@ export default class Query {
       .query(query)
       .timerange(timerange)
       .filter(filter)
+      .filters(filters)
       .searchTypes(searchTypes);
 
     return filter ? builder.filter(filter) : builder;
@@ -179,6 +196,7 @@ export default class Query {
       || !isDeepEqual(this.query, other.query)
       || !isDeepEqual(this.timerange, other.timerange)
       || !((isNullish(this.filter) && isNullish(other.filter)) || isDeepEqual(this.filter, other.filter))
+      || !((isNullish(this.filters) && isNullish(other.filters)) || isDeepEqual(this.filters, other.filters))
       || !isDeepEqual(this.searchTypes, other.searchTypes)) {
       return false;
     }
@@ -187,13 +205,14 @@ export default class Query {
   }
 
   toJSON(): QueryJson {
-    const { id, query, timerange, filter, searchTypes } = this._value;
+    const { id, query, timerange, filter, filters, searchTypes } = this._value;
 
     return {
       id,
       query,
       timerange,
       filter,
+      filters,
       search_types: searchTypes,
     };
   }
@@ -206,9 +225,9 @@ export default class Query {
   }
 
   static fromJSON(value: QueryJson): Query {
-    const { id, query, timerange, filter, search_types } = value;
+    const { id, query, timerange, filter, filters, search_types } = value;
 
-    return new Query(id, query, timerange, Immutable.fromJS(filter), search_types);
+    return new Query(id, query, timerange, Immutable.fromJS(filter), search_types, filters ? Immutable.List(filters) : filters);
   }
 }
 
@@ -239,13 +258,17 @@ class Builder {
     return new Builder(this.value.set('filter', Immutable.fromJS(value)));
   }
 
+  filters(value: FiltersType | null | undefined): Builder {
+    return new Builder(this.value.set('filters', value ? Immutable.List(value) : value));
+  }
+
   searchTypes(value: SearchTypeList): Builder {
     return new Builder(this.value.set('searchTypes', value));
   }
 
   build(): Query {
-    const { id, query, timerange, filter, searchTypes } = this.value.toObject();
+    const { id, query, timerange, filter, filters, searchTypes } = this.value.toObject();
 
-    return new Query(id, query, timerange, filter, searchTypes);
+    return new Query(id, query, timerange, filter, searchTypes, filters);
   }
 }
